requirements.txt
# General:
# deepspeed
# dlhpcstarter @ git+https://github.com/csiro-mlai/dl_hpc_starter_pack@d092644
duckdb
gdown
libauc
lightning
lmdb
monai
pandas
peft
pydantic
sentencepiece
snakeviz
tensorboard
# torch==2.1.*
torchmetrics
torch-summary
# torchvision==0.16.*
transformers

# Monitoring:
# neptune-tensorboard

# Project specific:
timm

# Demonstration:
streamlit

# End-of-life:
Pillow

# Metrics:
bert-score
pycocoevalcap
pycocotools
radgraph

# Results:
pingouin
plotnine
pypdf